web: refactoring to reduce number of dependencies
Web container now starts and display!
