+++ lib/portscanner_utils.py
@@ -1,0 +1,35 @@
+PORTS = { 8333  : "bitcoin",
+          9051  : "bitcoin-control",
+          9333  : "litecoin",
+          22556 : "dogecoin",
+          6697  : "irc",
+          6667  : "irc",
+          143   : "imap",
+          110   : "pop3",
+          119   : "nntp",
+          22    : "ssh",
+          2222  : "ssh?",
+          23    : "telnet",
+          25    : "smtp",
+          80    : "http",
+          443   : "https",
+          21    : "ftp",
+          5900  : "vnc",
+          27017 : "mongodb",
+          9200  : "elasticsearch",
+          3128  : "squid-proxy?",
+          8080  : "proxy?" ,
+          8118  : "proxy?" ,
+          8000  : "proxy?" ,
+          9878  : "richochet",
+          666   : "hail satan!",
+          31337 : "eleet",
+          1337  : "eleet",
+          69    : "good times",
+          6969  : "double the fun",
+          1234  : "patterns rule",
+          12345 : "patterns rule",
+        }
+
+def get_service_name(port):
+    return PORTS.get(port)
